- forecast fetch is made as Single instead of Observable - fixed types in data classes

+++ app/src/main/java/com/onos/weather/weatherapp/network/response/CurrentWeatherResponse.kt
@@ -15,13 +15,13 @@ data class CurrentWeatherResponse(val coord: Coord,
                                   val name: String,
                                   val cod: Int) {
 
-    data class Wind(val speed: Double, val deg: Int)
+    data class Wind(val speed: Double, val deg: Double)
 
     data class Clouds(val all: Int)
 
     data class Coord(val lon: Double, val lat: Double)
 
-    data class Main(val temp: Double, val pressure: Int, val humidity: Int,
+    data class Main(val temp: Double, val pressure: Double, val humidity: Int,
                     @SerializedName("temp_min") val tempMin: Double,
                     @SerializedName("temp_max") val tempMax: Double)
